# Code improvements of the Kerl binary-ternary conversion