… kerl/converter_test.go
… trinary/trinary.go