# Description of change

This PR adds code improvements and optimizations to the binary-ternary conversion used for Kerl:
- Add idiomatic go bigint package based on `[]uint32`
- Add more and better testcases for the exposed functions
- Add more comments to explain the math behind the conversion
- Use the `bits` package for better pattern matching support of CPU functionality
- Avoid slice allocations on the heap
- Reduce go bounds checks added by the compiler
- Unroll and inline hot functions

On the tested machine this leads to a significant speedup for all Kerl related functions. This has been tested and verified with a few artificial benchmarks:
### `amd64`
```
name                  old time/op  new time/op  delta
KerlTrytesToBytes-48   693ns ± 5%   492ns ± 7%  -29.06%
KerlBytesToTrytes-48   881ns ± 3%   533ns ± 2%  -39.50%
Kerl-48               2.88µs ± 5%  2.44µs ± 6%  -15.24%
KerlTrytes-48         2.85µs ± 6%  2.27µs ± 5%  -20.46%
```

### `386`
```
name                  old time/op  new time/op  delta
KerlTrytesToBytes-48   911ns ± 2%   604ns ± 2%  -33.75%
KerlBytesToTrytes-48  4.62µs ± 2%  1.45µs ± 3%  -68.60%
Kerl-48               8.66µs ± 4%  5.28µs ± 3%  -38.98%
KerlTrytes-48         8.82µs ± 4%  5.14µs ± 4%  -41.72%
```

## Type of change

- Enhancement (a non-breaking change which adds functionality)

## How the change has been tested

- Added ginko tests for `KerlBytesZeroLastTrit`
- Added converter tests for the edge case around the criticial ±⌊3²⁴² / 2⌋ values
- Artificial tests against the reference implementation

## Change checklist

- [x] My code follows the contribution guidelines for this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have added tests using ginkgo that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes